Refactor to avoid buffering, and fix sub-chunk hashes
It now pushes the data directly into an MD4 hasher, while keeping track
of when the chunk is done, to grab its hash and reset the hasher.

This takes significantly less memory and allocations, and is a little
bit faster (at least according to running the benchmarks on my PC).

It also ended up being a fix for the hashing of small inputs.